\name{uspop}
\alias{uspop}
\docType{data}
\title{United States Population Grid from the cart distribution}
\description{
 This is a table of 512 by 1024 values giving the US population
 on a grid. This comes from Mark Newman's cart distribution.
}
\usage{data(uspop)}
\format{
  This is data frame with dimensions 512 x 1024.

}
\source{
 Mark Newman's cart distribution.
 See \url{http://www-personal.umich.edu/~mejn/cart/}
}
\examples{
data(uspop)
}
\keyword{datasets}
